Revert "task 19 , 20"
