+++ src/app/category/khatamkari/Khatamkari.module.css
@@ -6,87 +6,6 @@
   text-align: center;
   color: var(--primary-color);
 }
-.loading {
-  text-align: center;
-  padding: 60px;
-  font-size: 18px;
-  color: #666;
-}
-/* ——— استایل‌های Collections ——— */
-.row {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 15px;
-  margin: 40px 0px;
-}
-
-/* ——— DESKTOP: 6 cards per row ——— */
-.collectionCard {
-  position: relative;
-  width: calc((100% - 5 * 15px) / 6);
-  aspect-ratio: 0.8;
-  border-radius: 8px;
-  text-decoration: none;
-  transition: transform 0.3s ease;
-  overflow: hidden; /* ensures image & overlay respect rounded corners */
-}
-
-.collectionCard:hover {
-  transform: scale(1.03);
-}
-
-.collectionImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover; /* keeps aspect ratio, fills card */
-  border-radius: 8px;
-  display: block;
-}
-
-/* Overlay & text */
-.overlay {
-  position: absolute;
-  inset: 0;
-  background: var(--primary-color-dark);
-  border-radius: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.description {
-  position: absolute;
-  bottom: 0px;
-  color: #fff;
-  font-size: 18px;
-  font-weight: 600;
-  padding: 10px;
-  text-align: center;
-}
-
-/* ——— TABLET: 4 cards per row ——— */
-@media (max-width: 1024px) {
-  .collectionCard {
-    width: calc((100% - 3 * 15px) / 4);
-    aspect-ratio: 0.8;
-  }
-}
-
-/* ——— MOBILE: 3 cards per row ——— */
-@media (max-width: 768px) {
-  .collectionCard {
-    width: calc((100% - 2 * 15px) / 3);
-    aspect-ratio: 0.8;
-  }
-
-  .description {
-    font-size: 12px;
-  }
-}
-
-
-
-
 
 .seoSection {
   direction: rtl;
