--- src/app/category/khatamkari/saatkhatamkari/Saatkhatamkari.module.css
@@ -1,85 +1,0 @@
-.pageContainer {
-    direction: rtl;
-  }
-  
-  .title {
-    text-align: center;
-    color: var(--primary-color);
-  }
-  
-  .seoSection {
-    direction: rtl;
-    text-align: right;
-    /* max-width: 1200px; */
-    margin: 40px auto;
-    padding: 20px;
-    line-height: 1.8;
-    font-size: 1rem;
-    color: #333;
-    background-color: #fafafa;
-    border-radius: 8px;
-  }
-  
-  .seoSection h2 {
-    font-size: 1.4rem;
-    margin: 32px 0 16px;
-    font-weight: 700;
-    color: #222;
-    border-right: 4px solid var(--primary-color, #c49a6c);
-    padding-right: 8px;
-  }
-  
-  .seoSection h3 {
-    font-size: 1.2rem;
-    margin: 24px 0 12px;
-    font-weight: 600;
-    color: #444;
-  }
-  
-  .seoSection p {
-    margin-bottom: 16px;
-    color: #555;
-  }
-  
-  .seoSection ul,
-  .seoSection ol {
-    margin: 16px 0 24px;
-    padding-right: 20px;
-  }
-  
-  .seoSection ul li,
-  .seoSection ol li {
-    margin-bottom: 10px;
-    line-height: 1.7;
-    color: #444;
-  }
-  
-  .seoSection strong {
-    color: var(--primary-color, #c49a6c);
-  }
-  
-  .seoSection img {
-    max-width: 100%;
-    height: auto;
-    margin: 20px 0;
-    border-radius: 6px;
-  }
-  
-  @media (max-width: 768px) {
-    .pageContainer {
-      margin-top: 60px;
-    }
-    .seoSection {
-      padding: 15px;
-      font-size: 0.95rem;
-    }
-  
-    .seoSection h2 {
-      font-size: 1.2rem;
-    }
-  
-    .seoSection h3 {
-      font-size: 1.05rem;
-    }
-  }
-
