--- src/app/category/khatamkari/saatkhatamkari/page.js
@@ -1,135 +1,0 @@
-import Header from "@/components/Header/Header";
-import Footer from "@/components/Footer/Footer";
-import CategoryClient from "@/components/CategoryClient/CategoryClient";
-import styles from "./Saatkhatamkari.module.css";
-import FooterMenu from "@/components/FooterMenu/FooterMenu";
-
-const API_URL = "https://api.kimiatoranj.com/";
-
-export const metadata = {
-  title: "خرید ساعت خاتم‌کاری اصفهان | انواع ساعت دیواری و رومیزی خاتم دست‌ساز",
-  description:
-    "ساعت خاتم‌کاری اصفهان، تلفیق هنر و زمان. خرید انواع ساعت دیواری و رومیزی خاتم‌کاری دست‌ساز با طرح‌های اصیل ایرانی، مناسب هدیه و دکوراسیون شیک. فروشگاه کیمیا ترنج.",
-};
-
-export default async function KhatamkariPage() {
-  const productsRes = await fetch(
-    `${API_URL}api/store/products/?collection=خاتم کاری&page=1`,
-    { next: { revalidate: 60 } }
-  );
-
-  const productsData = productsRes.ok
-    ? await productsRes.json()
-    : { results: [], next: null };
-
-  return (
-    <>
-      <Header />
-
-      <div className={styles.pageContainer}>
-        {/* Product Grid */}
-        <h1 className={styles.title}>محصولات ساعت خاتم کاری کیمیاترنج </h1>
-        <CategoryClient
-          categoryName="خاتم کاری"
-          initialProducts={productsData.results}
-          initialHasMore={!!productsData.next}
-        />
-        <div className={styles.seoSection}>
-          <h1>خرید ساعت خاتم کاری اصفهان </h1>
-
-          <p>
-            ساعت خاتم‌کاری یکی از خاص‌ترین صنایع دستی اصفهان است که نه‌تنها
-            کاربردی برای نمایش زمان دارد، بلکه به‌عنوان یک اثر هنری اصیل، زیبایی
-            و شکوه دکوراسیون شما را دوچندان می‌کند. این ساعت‌ها حاصل تلفیق چوب،
-            استخوان و فلز با طرح‌های هندسی ظریف هستند که توسط استادکاران اصفهانی
-            با دقت و صبر ساخته می‌شوند.
-          </p>
-
-          <h2>ساعت خاتم‌کاری چیست؟ </h2>
-
-          <p>
-            خاتم‌کاری یکی از هنرهای اصیل ایرانی است که در آن قطعات بسیار ریز
-            چوب، فلز و استخوان در کنار هم قرار می‌گیرند تا اشکال هندسی منظم خلق
-            شود. وقتی این هنر در ساخت ساعت به کار می‌رود، محصولی منحصربه‌فرد به
-            دست می‌آید؛ ساعتی که علاوه بر نمایش زمان، روایتگر تاریخ و فرهنگ
-            ایران زمین است.
-          </p>
-          <h2>انواع ساعت خاتم‌کاری </h2>
-          <h3>ساعت دیواری خاتم‌کاری </h3>
-          <p>
-            این مدل از محبوب‌ترین انتخاب‌ها برای خانه‌های ایرانی است. ساعت
-            دیواری خاتم‌کاری با طرح‌های متنوع، جلوه‌ای باشکوه به پذیرایی یا دفتر
-            کار شما می‌دهد.
-          </p>
-
-          <h3>ساعت رومیزی خاتم‌کاری </h3>
-          <p>
-            اگر به دنبال هدیه‌ای ظریف و ارزشمند هستید، ساعت رومیزی خاتم‌کاری
-            انتخابی ایده‌آل است. این محصولات علاوه بر زیبایی، اندازه جمع‌وجوری
-            دارند و برای میز کار یا قفسه‌ها بسیار مناسب‌اند.
-          </p>
-
-          <h3>ساعت تزئینی خاتم </h3>
-          <p>
-            بعضی از ساعت‌ها جنبه کاربردی کمتری دارند و بیشتر به‌عنوان یک اثر
-            هنری و دکوری شناخته می‌شوند. این مدل‌ها معمولاً برای کلکسیون‌داران و
-            علاقه‌مندان به هنر ایرانی ارزش زیادی دارند.
-          </p>
-
-          <h2>ویژگی‌ها و مزایای ساعت خاتم‌کاری </h2>
-
-          <ul>
-            <li>ساخته دست استادکاران اصفهانی</li>
-            <li>ترکیب هنر و کارایی</li>
-            <li>دارای ارزش مادی و معنوی بالا</li>
-            <li>کیفیت ساخت عالی و ماندگاری طولانی</li>
-            <li>انتخابی خاص برای هدیه در مناسبت‌های ویژه</li>
-          </ul>
-
-          <h2>کاربرد ساعت خاتم‌کاری در دکوراسیون داخلی </h2>
-
-          <p>
-            یک ساعت خاتم می‌تواند نقطه کانونی اتاق شما باشد. استفاده از آن در
-            پذیرایی، دفتر کار یا حتی اتاق خواب، حس اصالت و آرامش را منتقل
-            می‌کند. ترکیب رنگ‌های طبیعی چوب و طرح‌های هندسی خاتم، هماهنگی
-            چشم‌نوازی با سبک‌های مختلف دکوراسیون ایجاد می‌کند.
-          </p>
-          <h2>چرا ساعت خاتم‌کاری هدیه‌ای خاص است؟ </h2>
-          <p>
-            هدیه دادن ساعت خاتم‌کاری، به‌ویژه در مناسبت‌هایی مثل سالگرد ازدواج،
-            روز پدر یا مادر و جشن‌های خانوادگی، نشانه‌ای از احترام و علاقه است.
-            این هدیه نه‌تنها کاربردی است، بلکه ارزش هنری و فرهنگی بالایی دارد و
-            می‌تواند سال‌ها در خانه عزیزانتان ماندگار بماند.
-          </p>
-          <h2>خرید اینترنتی ساعت خاتم‌کاری از کیمیا ترنج </h2>
-          <p>
-            فروشگاه کیمیا ترنج با سال‌ها تجربه در عرضه صنایع دستی اصیل اصفهان،
-            مجموعه‌ای متنوع از ساعت‌های خاتم را در طرح‌ها و اندازه‌های مختلف
-            ارائه می‌دهد. در این فروشگاه می‌توانید ساعت دیواری، رومیزی و تزئینی
-            خاتم‌کاری را با ضمانت اصالت و کیفیت تهیه کنید.
-          </p>
-
-          <h2>مزایای خرید ساعت خاتم کاری از کیمیا ترنج </h2>
-
-          <ul>
-            <li>عرضه مستقیم از کارگاه‌های خاتم‌کاری اصفهان</li>
-            <li>تنوع بالا در مدل‌ها و طرح‌ها</li>
-            <li>قیمت مناسب و بدون واسطه</li>
-            <li>ارسال مطمئن به سراسر کشور </li>
-          </ul>
-
-          <h2>و در آخر ... </h2>
-
-          <p>
-            ساعت خاتم‌کاری فراتر از یک وسیله برای نشان دادن زمان است؛ این محصول
-            نمادی از اصالت، فرهنگ و هنر ایرانی به شمار می‌رود. اگر به دنبال
-            هدیه‌ای ماندگار یا دکوری شیک برای خانه خود هستید، ساعت خاتم‌کاری
-            انتخابی بی‌نظیر است.
-          </p>
-        </div>
-      </div>
-      <FooterMenu />
-      <Footer />
-    </>
-  );
-}
