--- src/app/category/khatamkari/shokolatkhori/page.js
@@ -1,131 +1,0 @@
-import Header from "@/components/Header/Header";
-import Footer from "@/components/Footer/Footer";
-import CategoryClient from "@/components/CategoryClient/CategoryClient";
-import styles from "./Shokolatkhori.module.css";
-import FooterMenu from "@/components/FooterMenu/FooterMenu";
-
-const API_URL = "https://api.kimiatoranj.com/";
-
-export const metadata = {
-  title: "خرید شکلات‌خوری خاتم‌کاری اصفهان | انواع شکلات‌خوری دست‌ساز خاتم",
-  description:
-    "انواع شکلات‌خوری خاتم‌کاری اصفهان دست‌ساز؛ انتخابی شیک برای پذیرایی و هدیه. خرید اینترنتی شکلات‌خوری خاتم با طرح‌های اصیل ایرانی از فروشگاه کیمیا ترنج.",
-};
-
-export default async function KhatamkariPage() {
-  const productsRes = await fetch(
-    `${API_URL}api/store/products/?collection=خاتم کاری&page=1`,
-    { next: { revalidate: 60 } }
-  );
-
-  const productsData = productsRes.ok
-    ? await productsRes.json()
-    : { results: [], next: null };
-
-  return (
-    <>
-      <Header />
-
-      <div className={styles.pageContainer}>
-        {/* Product Grid */}
-        <h1 className={styles.title}>
-          محصولات شکلات خوری خاتم کاری کیمیاترنج{" "}
-        </h1>
-        <CategoryClient
-          categoryName="خاتم کاری"
-          initialProducts={productsData.results}
-          initialHasMore={!!productsData.next}
-        />
-        <div className={styles.seoSection}>
-          <h1>خرید شکلات‌خوری خاتم‌کاری اصفهان </h1>
-
-          <p>
-            شکلات‌خوری خاتم‌کاری یکی از اصیل‌ترین و زیباترین محصولات صنایع دستی
-            اصفهان است که علاوه بر کاربرد در پذیرایی، جلوه‌ای خاص و فاخر به
-            دکوراسیون منزل می‌دهد. این ظروف دست‌ساز با ترکیب چوب، فلز و استخوان
-            در قالب نقش‌های هندسی منظم ساخته می‌شوند و نتیجه آن، محصولی ارزشمند
-            و ماندگار است.
-          </p>
-
-          <h2>شکلات‌خوری خاتم‌کاری چیست؟ </h2>
-
-          <p>
-            خاتم‌کاری هنری ظریف و کهن است که در آن قطعات ریز چوب، استخوان و فلز
-            کنار هم قرار می‌گیرند تا طرح‌هایی چشم‌نواز شکل بگیرد. وقتی این هنر
-            در ساخت شکلات‌خوری به کار می‌رود، حاصل آن ظرفی است که هم جنبه تزئینی
-            دارد و هم کاربردی.
-          </p>
-          <h2>انواع شکلات‌خوری خاتم‌کاری </h2>
-          <h3>شکلات‌خوری خاتم پایه‌دار </h3>
-          <p>
-            امدلی شیک و مجلل که بیشتر برای پذیرایی رسمی استفاده می‌شود. پایه
-            خاتم‌کاری‌شده جلوه‌ای باشکوه به میز پذیرایی شما می‌دهد.
-          </p>
-
-          <h3>شکلات‌خوری خاتم بدون پایه </h3>
-          <p>
-            ساده‌تر و کاربردی‌تر، مناسب استفاده روزمره یا پذیرایی‌های خانوادگی.
-          </p>
-
-          <h3>شکلات‌خوری ترکیبی (خاتم و مس یا خاتم و مینا) </h3>
-          <p>
-            باین مدل‌ها با ترکیب چند هنر صنایع دستی ساخته می‌شوند و جلوه‌ای
-            بسیار خاص و متفاوت دارند.
-          </p>
-
-          <h2>ویژگی‌ها و مزایای شکلات‌خوری خاتم‌کاری </h2>
-
-          <ul>
-            <li>ساخت دست‌ساز توسط استادکاران اصفهانی</li>
-            <li>دوام و ماندگاری بالا</li>
-            <li>جلوه‌ای خاص برای پذیرایی رسمی و مهمانی‌ها</li>
-            <li>قابلیت استفاده به‌عنوان هدیه‌ای ارزشمند</li>
-            <li>
-              هماهنگی با سایر ظروف صنایع دستی مانند گلدان یا سینی خاتم‌کاری
-            </li>
-          </ul>
-
-          <h2>کاربرد شکلات‌خوری خاتم در دکوراسیون داخلی خانه </h2>
-
-          <p>
-            یشکلات‌خوری خاتم‌کاری تنها یک ظرف پذیرایی نیست، بلکه یک عنصر تزئینی
-            برای دکوراسیون منزل هم به شمار می‌رود. قرار دادن آن روی میز پذیرایی
-            یا ویترین، اصالت و هنر ایرانی را در خانه به نمایش می‌گذارد.
-          </p>
-          <h2>چرا شکلات‌خوری خاتم‌کاری گزینه ای مناسب برای هدیه است؟ </h2>
-          <p>
-            این محصولات به دلیل ارزش هنری بالا و طراحی منحصربه‌فرد، گزینه‌ای
-            فوق‌العاده برای هدیه دادن در مناسبت‌هایی چون عروسی، سالگرد ازدواج یا
-            مراسم رسمی هستند. هدیه‌ای که هم زیباست و هم ماندگار.
-          </p>
-          <h2>خرید اینترنتی شکلات‌خوری خاتم‌کاری از کیمیا ترنج </h2>
-          <p>
-            در فروشگاه کیمیا ترنج مجموعه‌ای متنوع از شکلات‌خوری‌های خاتم‌کاری
-            دست‌ساز اصفهان عرضه می‌شود. شما می‌توانید مدل‌های مختلف را بررسی
-            کنید و با خرید مستقیم از هنرمندان، از اصالت و کیفیت محصول مطمئن
-            باشید.
-          </p>
-
-          <h2>مزایای خرید شکلات خوری خاتم کاری از کیمیا ترنج </h2>
-
-          <ul>
-            <li>تنوع بالا در طرح و مدل</li>
-            <li>قیمت مناسب و بدون واسطه</li>
-            <li>ضمانت اصالت صنایع دستی</li>
-            <li>ارسال سریع و مطمئن به سراسر کشور</li>
-          </ul>
-
-          <h2>و در آخر ... </h2>
-
-          <p>
-            شکلات‌خوری خاتم‌کاری نمادی از اصالت هنر ایرانی است. چه به دنبال ظرفی
-            خاص برای پذیرایی باشید و چه هدیه‌ای ماندگار برای عزیزانتان، این
-            محصول انتخابی ارزشمند خواهد بود.
-          </p>
-        </div>
-      </div>
-      <FooterMenu />
-      <Footer />
-    </>
-  );
-}
